requirements.txt
Flask==3.1.0
peewee==3.17.8
pydantic_settings==2.7.1
python_bcrypt==0.3.2
